chore: Update Version API
This change modifies the workflow for updating the API version. It updates the git configuration for using the GITHUB_ACTOR variable, and adds and commits the changes in the working tree. The commit message is set to "chore: Update Version API". It also checks if there are changes to be committed, and if not, the workflow exits with the message "Nothing to commit, working tree clean.". Finally, it pushes the files to GitHub and deploys the environment if there were any changes committed.